// client/src/app/api/models/chat-message-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { JsonNode } from '../models/json-node';
import { MessageRole } from '../models/message-role';
export interface ChatMessageResponse {
  createdAt: string;
  nanoid: string;
  parts?: Array<JsonNode> | null;
  producedVersionNanoid?: string | null;
  role: MessageRole;
  text: string;
}